+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Changed
 - Exponential operations merged with hyperbolic functions
 - Space is replaced by Group. Space is reserved for vector spaces.
+- VectorSpace is now a vector space
 
 ### Deprecated